No need to duplicate matrices
While we were doing multiple colors, we did construct our own matrices,
but in retrospect, once the palette was reduced to R, G, and B, the
least square matrix turned into the XYZ to linear sRGB matrix :). We
just realized this and have adjusted to code to not declare it again.

In retrospect, it seems so obvious.