Rework approach of mixing to match latest Spectral.js (#35)
The actual implementation works exactly the same, but it is calculated
a bit different. Instead of adjusting the progress and then applying
lerp on each KS, the calculations are reworked to calculate mixing
concentrations. Then each KS can just be multiplied by the concentration
and summed with the other colors KS and concentration. This follows more
closely to what is being done in Spectral.js 3+.